+++ rpc/src/v1/traits/traces.rs
@@ -28,7 +28,7 @@ build_rpc_trait! {
 
 		/// Returns traces matching given filter.
 		#[rpc(name = "trace_filter")]
-		fn filter(&self, TraceFilter, Option<TraceIndex>, Option<u64>) -> Result<Option<Vec<LocalizedTrace>>, Error>;
+		fn filter(&self, TraceFilter, Option<TraceIndex>, Option<usize>) -> Result<Option<Vec<LocalizedTrace>>, Error>;
 
 		/// Returns transaction trace at given index.
 		#[rpc(name = "trace_get")]
